Updating to latest FlutterFlow output.

+++ lib/custom_code/actions/add_to_hold_list_gr_calculationdis_per.dart
@@ -19,11 +19,11 @@ Future<List<dynamic>> addToHoldListGrCalculationdisPer(
   List<TaxMasterRecord> taxcollection,
   String inclusiveorexclusive,
   List<UnitTypeRecord> unitcollection,
-  int disPer,
+  double disPer,
   double disAmt,
   double ratePrice,
   double qty,
-  int taxPer,
+  double taxPer,
   double taxAmt,
 ) async {
   List<dynamic> list = FFAppState().allBillsList;
