+++ lib/custom_code/actions/dis_amt_changed_functiongrocery.dart
@@ -20,11 +20,11 @@ Future<List<dynamic>> disAmtChangedFunctiongrocery(
   List<TaxMasterRecord> taxcollection,
   String inclusiveorexclusive,
   List<UnitTypeRecord> unitcollection,
-  int disPer,
+  double disPer,
   double disAmt,
   double ratePrice,
   double qty,
-  int taxPer,
+  double taxPer,
   double taxAmt,
 ) async {
   List<dynamic> list = FFAppState().allBillsList;
