+++ lib/custom_code/actions/tax_per_changed_functiongrocery.dart
@@ -18,11 +18,11 @@ Future<List<dynamic>> taxPerChangedFunctiongrocery(
   List<TaxMasterRecord> taxcollection,
   String inclusiveorexclusive,
   List<UnitTypeRecord> unitcollection,
-  int disPer,
+  double disPer,
   double disAmt,
   double ratePrice,
   double qty,
-  int taxPer,
+  double taxPer,
   double taxAmt,
 ) async {
   List<dynamic> list = FFAppState().allBillsList;
